+++ src/MotionSequence.h
@@ -26,13 +26,13 @@ enum class MotionSequenceStatusType
 class Stage
 {
 public:
-    int stageId() { return _stageId; }
-    int motorId() { return motor->getId(); }
-    int goalPosition() { return _goalPosition; }
-    int accuracy() { return _accuracy; }
+    int8_t stageId() { return _stageId; }
+    uint8_t motorId() { return motor->getId(); }
+    int32_t goalPosition() { return _goalPosition; }
+    int32_t accuracy() { return _accuracy; }
     bool started() { return _started; }
 
-    void update(int stageId, DXLMotor *motor, int goalPosition, int accuracy);
+    void update(int8_t stageId, DXLMotor *motor, int32_t goalPosition, int32_t accuracy);
 
     /**
      * @return -1: error (position goal is not the same as stage goal)
@@ -59,19 +59,19 @@ public:
 
 private:
     bool _started;
-    int _stageId = -1;
+    int8_t _stageId = -1;
     DXLMotor *motor = nullptr;
-    int _goalPosition = -1;
-    int _accuracy = -1;
+    int32_t _goalPosition = -1;
+    int32_t _accuracy = -1;
 };
 
 class MotionSequence
 {
 public:
     MotionSequence(){};
-    MotionSequence(MotionSequenceType type, DXLMotor *motors[], const int sequence[]);
+    MotionSequence(MotionSequenceType type, DXLMotor *motors[], const uint16_t sequence[]);
     MotionSequenceType sequenceType() { return _sequenceType; }
-    int currentStageId() { return currentStage.stageId(); }
+    int8_t currentStageId() { return currentStage.stageId(); }
     MotionSequenceStatusType status();
 
     /**
@@ -86,7 +86,7 @@ public:
 private:
     MotionSequenceType _sequenceType = MotionSequenceType::UNKNOWN;
     DXLMotor **motors;
-    const int *sequence;
+    const uint16_t *sequence;
     Stage currentStage;
 };
 #endif
